Streamlined the 'legacy' graphics code implementation.
The vtable-style code that was implementing the drawing of primitives
turned out to be entirely unnecessary; every virtual call was inside a
switch or if on the type of the primitive.  Code has been restructured to
be more consistent and simpler.  Never-used primitives (composites and
polygons) have been removed entirely.


git-svn-id: svn://svn.code.sf.net/p/sc2/code/trunk@838 8092fc87-c524-0410-9efc-e669fe64eaf9