Improved save file format inspired by IFF.
Savefile is built of chunks that include their length. The only
two right now are Summ (the summary, must be first) and OmnZ
(the omnibus compressed data).

Padding that was originally used for alignment purposes has been
removed from the format, unless is it is part of the compressed
data.