Save format refactoring: Uncompressed save format
Legacy files are handled by load_legacy.c, but this now should be
the only part of the code that still uses the custom decompression
library.

These save formats are very much interim and held for testing
purposes! Do not use these commits for actual play, or your saves
will be lost.