Reorganize the GameState array.
This is preparation work so that when the state files are removed
from the savegame format (or possibly removed entirely) it will
not be a savegame compatibility-breaking change.

Legacy support has been modified so that the GameState bits are
reordered as part of load, so loading legacy saves still works.