Fix the ZFP dialog desync bug introduced post 0.7.
The core issue is that the sound-advancement callbacks (used only
for altering the ZFP dialog locations) were altered to be launched
asynchronously from the main thread. This meant that new dialog
could be detected before the waiting callbacks had run.

This patch moves audio callbacks back into the audio decoding thread
and protects them with the speech-channel mutex. This should
guarantee that cur_sub_chunk only appears as a value when consulting
the current subtitles after the callback has run completely.

Logic has also been added to CheckSubtitles to raise a warning in
the logs if the values get out of sync.