Fix two bugs regarding stale SIS messages, from Kruzen
When entering Hyperspace on Auto-Pilot, or are intercepted by
neutral or hostile alien ships during Auto-Pilot, the text at the
top of the screen doesn't update properly until the status
would need to change. This corrects the times when DrawSisMessage
is called to keep that up-to-date as needed.

This patch was backported from the Mega-Mod.